Fix file splits
